requirements.txt
absolute-import==0.4
amqp==5.2.0
anyio==4.6.2.post1
arrow==1.3.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-yasg==1.21.8
future==1.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
inflection==0.5.1
inform==1.31
kombu==5.4.2
packaging==24.1
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg==3.2.3
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pytz==2024.2
PyYAML==6.0.2
quantiphy==2.20
redis==5.1.1
requests==2.32.3
schedule-reminder==1.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
utils==1.0.2
vine==5.1.0
wcwidth==0.2.13
